pyproject.toml
[project]
name = "mcp-ops-ai-use-monitor"
version = "0.1.0"
description = "MCP Server for monitoring Claude Code usage with Langfuse integration"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.9.0",
    "langfuse>=3.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
